💻 Now separated app and index in two different file for the backend
